backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
opencv-python-headless>=4.8.0
numpy>=1.21.0
Pillow>=9.0.0
python-dotenv>=0.19.0
aiofiles>=0.8.0
